requirements.txt
eel
pyserial